+++ src/main/java/zmaster587/advancedRocketry/world/GenLayerBiomePlanet.java
@@ -1,21 +1,13 @@
 package zmaster587.advancedRocketry.world;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
-import java.util.Random;
-import java.util.logging.Logger;
 
-import zmaster587.advancedRocketry.AdvancedRocketry;
-import cpw.mods.fml.common.FMLCommonHandler;
-import net.minecraft.util.MathHelper;
 import net.minecraft.util.WeightedRandom;
 import net.minecraft.world.WorldType;
 import net.minecraft.world.biome.BiomeGenBase;
 import net.minecraft.world.gen.layer.GenLayer;
 import net.minecraft.world.gen.layer.IntCache;
-import net.minecraftforge.common.BiomeDictionary;
-import net.minecraftforge.common.BiomeManager;
 import net.minecraftforge.common.BiomeManager.BiomeEntry;
 
 public class GenLayerBiomePlanet extends GenLayer {
@@ -54,7 +46,7 @@ public class GenLayerBiomePlanet extends GenLayer {
 		{
 			for (int j1 = 0; j1 < p_75904_3_; ++j1)
 			{
-				this.initChunkSeed((long)(j1 + p_75904_1_), (long)(i1 + p_75904_2_));
+				this.initChunkSeed(j1 + p_75904_1_, i1 + p_75904_2_);
 				
 				aint1[j1 + i1 * p_75904_3_] = getWeightedBiomeEntry().biome.biomeID;
